requirements.txt
Pillow==10.1.0
customtkinter==5.2.2
tkinterdnd2==0.3.0 